Split store into multiple files for easier management, and make auth user available site wide
