+++ TODO.md
@@ -1,0 +1,10 @@
+# Switch to Vite
+# Switch to Pinia
+# Use TailwindCSS for styling
+# Switch to Composition API
+# Create dedicated Thread component
+# Create dedicated Forum component
+# Create dedicated Post component
+# Convert to TypeScript
+# Add tests
+# Upgrade to using firebase v9
